Adds instructions for Windows
vlab can be compiled and run on Windows 10 and 11 using WSL. Version 1 of WSL works but it is better to use WSL 2 because it includes an Xserver. Some compilations issues may occur with newer versions of gcc on Ubuntu 22, e.g., Rayshade may need to be disabled. This needs to be investigated.