Adds user warning before deleting vlab's temp folder
When browser starts it will check if there are object files in the temp folder (lab table). If it is not empty, the user will be prompted on what to do. The name of the folder in the temp folder was updated to make more clear which objects are on the lab table.