Fixes saving images with alpha channel in cpfg
The menu option for enabling alpha channel was not set for PDF, PNG and TIF formats in cpfg. It looks like the SaveAs dialog windows was copied from lpfg, but, because the image formats are not the same, the index for enabling the alpha channel combo box was wrong. I added a variable to set the index but the whole solution should be more robust.