Fixes texture loading in vv
Texture loading in vv was causing a seg fault because the Texture2D constructor was making OpenGL calls without a valid context. The constructor is no longer available but was replaced by a new load(filename) function. In addition, since texture loading is implemented with QImage, the vv model had to be compiled against Qt libraries. Changed the vvlib from static to dynamic so that it would not be necessary. But this probably works on mac only, the QMAKE_SONAME_PREFIX should be changed for GNU/Linux.