Fixes lpfg window redrawing on Retina displays
Detects if the lpfg window is moved between two screens with different device pixel ratios, and, if so, repaints the window with the new ratio. The trick was to recompute the projection matrix before repainting. I think the other programs (besides lpfg) must be recomputing the projection on each paint call--they don't seem to have the same issue.