Removes mouse x,y bounds check in snapicon
On some display arrangements on macOS, the bounds check causes snapicon to be stuck on the primary screen. This is because the primary screen could be below the secondary screen and the x,y position of snapicon can be negative. A perfect solution would be to save the min and max of the screen extents and compare to those but I don't see a way to move  snapicon right off the screen.